requirements.txt
numpy
pandas
scikit-learn
tensorflow
nltk
polyaxon-cli
google-cloud-storage
